requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn>=0.29.0